backend/requirements.txt
aerich==0.6.2
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
databases==0.5.5
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.74.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
iso8601==0.1.16
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
tomlkit==0.10.0
tortoise-orm==0.18.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
